fix(#76): Resolve model dropdown issue in ConfigPanel settings
- Converted HTML string concatenation to template literals
- Fixed model dropdown disappearing when switching to Ollama provider
- Ensured Ollama models auto-populate in dropdown
- Resolved 'Unexpected end of input' syntax error in webview
- Improved overall template maintainability
