--- vscode-extension/jest.config.js
@@ -1,16 +1,0 @@
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  roots: ['<rootDir>/tests'],
-  testMatch: ['**/*.test.ts'],
-  moduleFileExtensions: ['ts', 'js'],
-  moduleNameMapper: {
-    '^vscode$': '<rootDir>/tests/__mocks__/vscode.ts'
-  },
-  collectCoverageFrom: [
-    'src/**/*.ts',
-    '!src/**/*.d.ts'
-  ],
-  coverageDirectory: 'coverage',
-  verbose: true
-};
