+++ vscode-extension/v1/.context/settings-persistence-fix.md
