+++ vscode-extension/v1/.gitignore
@@ -1,0 +1,8 @@
+node_modules/
+dist/
+*.vsix
+.DS_Store
+out/
+build/
+*.log
+.vscode-test/
