+++ vscode-extension/v1/src/panels/CodeSuggestionPanel.ts
