+++ vscode-extension/v1/src/panels/StatisticsPanel.ts
