+++ vscode-extension/v1/src/services/TraceService.ts
