+++ vscode-extension/v1/tests/integration/codeIntelligence.test.ts
