+++ vscode-extension/v1/tests/panels/CodeSuggestionPanel.test.ts
