+++ vscode-extension/v1/tests/services/MetricsService.test.ts
